core-java-modules/core-java-concurrency-basic/src/main/java/com/tom/comcurrent/completablefuture/CompletableFutureDemo.java
package com.tom.comcurrent.completablefuture;

import java.util.concurrent.CompletableFuture;
import java.util.concurrent.ExecutionException;
import java.util.concurrent.Executors;
import java.util.concurrent.Future;

/**
 * 功能描述
 *
 * @author TomLuo
 * @date 2020/2/15
 */
public class CompletableFutureDemo {
    public Future<String> calculateAsync() throws InterruptedException {
        CompletableFuture<String> completableFuture
                = new CompletableFuture<>();

        Executors.newCachedThreadPool().submit(() -> {
            Thread.sleep(500);
            completableFuture.complete("Hello");
            return null;
        });

        return completableFuture;
    }


    public Future<String> calculateAsyncWithCancellation() throws InterruptedException {
        CompletableFuture<String> completableFuture = new CompletableFuture<>();

        Executors.newCachedThreadPool().submit(() -> {
            Thread.sleep(500);
            //completableFuture.cancel(false)
            completableFuture.cancel(false);
            return null;
        });

        return completableFuture;
    }

}


